standard_data_format/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4
annotated-types==0.7.0
async-timeout==3.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cmd2==1.5.0
colorama==0.4.6
commonmark==0.9.1
et-xmlfile==2.0.0
filelock==3.17.0
filetype==1.2.0
fsspec==2025.2.0
ftfy==6.3.1
google-auth==2.38.0
google-genai==1.2.0
huggingface-hub==0.29.0
idna==2.10
jinja2==3.0.0
joblib==1.4.2
markdown2==2.5.3
markdownify==0.13.1
marker-pdf==1.5.3
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdftext==0.6.0
pillow==10.4.0
propcache==0.2.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pygments==2.19.1
pypdfium2==4.30.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
rapidfuzz==3.12.1
regex==2024.11.6
requests==2.32.3
rich==10.2.0
rsa==4.9
ruamel-yaml==0.17.17
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soupsieve==2.6
surya-ocr==0.11.1
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
triton==3.2.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
